Modified some requirements in requirements.txt to install only the latest version of the dependencies.
Added some checks in the code that verifies if the value in the .env file are not empty if there is empty fields it will raise an error.

Added a link to generate the refresh token instead of searching.
Added a warning in the console if not refresh rate was specified.

Removed unused imports.
Renamed some variables for the Spotify API to match the rest of the env variables.

Renamed sendMusicNote to send_music_note

Corrected some typos in the code.